ci: update reusable workflow versions
This is to get the fix in infrastructure due to the github checkout action causing issues.  There has been a bug introduced into the action, so Infrastructure has been updated from v3 to v4.  This way the workflows can work correctly so the prepare release and release workflows can be executed.