🚀Preview Release (v1.0.0-preview.26) (#727)
* start work for preview release

* release: updated version to v1.0.0-preview.26

* release: create release notes for version v1.0.0-preview.26

* release: improve release notes

* ide: add word to dictionary for vscode users

* ci: update workflow triggers

* ci: update reusable workflow versions

This is to get the fix in infrastructure due to the github checkout action causing issues.  There has been a bug introduced into the action, so Infrastructure has been updated from v3 to v4.  This way the workflows can work correctly so the prepare release and release workflows can be executed.